+++ src/FileSystemLoader.ts
@@ -5,26 +5,6 @@ import * as fs from 'fs'
 import * as path from 'path'
 import { Plugin } from "postcss";
 
-// Sorts dependencies in the following way:
-// AAA comes before AA and A
-// AB comes after AA and before A
-// All Bs come after all As
-// This ensures that the files are always returned in the following order:
-// - In the order they were required, except
-// - After all their dependencies
-const traceKeySorter = ( a: string, b: string ): number => {
-  if ( a.length < b.length ) {
-    return a < b.substring( 0, a.length ) ? -1 : 1
-  } else if ( a.length > b.length ) {
-    return a.substring( 0, b.length ) <= b ? -1 : 1
-  } else {
-    return a < b ? -1 : 1
-  }
-};
-
-export type Dictionary<T> = {
-  [key: string]: T | undefined;
-};
 
 export default class FileSystemLoader {
   private root: string;
@@ -96,3 +76,24 @@ export default class FileSystemLoader {
     return this;
   }
 }
+
+// Sorts dependencies in the following way:
+// AAA comes before AA and A
+// AB comes after AA and before A
+// All Bs come after all As
+// This ensures that the files are always returned in the following order:
+// - In the order they were required, except
+// - After all their dependencies
+function traceKeySorter( a: string, b: string ): number {
+  if ( a.length < b.length ) {
+    return a < b.substring( 0, a.length ) ? -1 : 1
+  } else if ( a.length > b.length ) {
+    return a.substring( 0, b.length ) <= b ? -1 : 1
+  } else {
+    return a < b ? -1 : 1
+  }
+};
+
+type Dictionary<T> = {
+  [key: string]: T | undefined;
+};
