+++ src/cli.ts
@@ -5,7 +5,7 @@ import * as chokidar from 'chokidar';
 import glob from 'glob';
 import * as yargs from 'yargs';
 import chalk from 'chalk';
-import {DtsCreator} from './dtsCreator';
+import {DtsCreator} from './DtsCreator';
 
 let yarg = yargs.usage('Create .css.d.ts from CSS modules *.css files.\nUsage: $0 [options] <input directory>')
   .example('$0 src/styles', '')
